Refactor isLargeText
