+++ packages/alfa-rules/src/common/predicate.ts
@@ -1,0 +1,27 @@
+export * from "./predicate/has-accessible-name";
+export * from "./predicate/has-attribute";
+export * from "./predicate/has-cascaded-value-declared-in-inline-style";
+export * from "./predicate/has-descendant";
+export * from "./predicate/has-explicit-role";
+export * from "./predicate/has-heading-level";
+export * from "./predicate/has-implicit-role";
+export * from "./predicate/has-inclusive-descendant";
+export * from "./predicate/has-inline-style-property";
+export * from "./predicate/has-non-empty-accessible-name";
+export * from "./predicate/has-outline";
+export * from "./predicate/has-role";
+export * from "./predicate/has-text-content";
+export * from "./predicate/has-text-decoration";
+export * from "./predicate/has-value";
+export * from "./predicate/is-document-element";
+export * from "./predicate/is-focusable";
+export * from "./predicate/is-inert";
+export * from "./predicate/is-keyboard-actionable";
+export * from "./predicate/is-large-text";
+export * from "./predicate/is-marked-decorative";
+export * from "./predicate/is-perceivable";
+export * from "./predicate/is-tabbable";
+export * from "./predicate/is-transparent";
+export * from "./predicate/is-visible";
+export * from "./predicate/is-whitespace";
+export * from "./predicate/reference-same-resource";
