+++ packages/alfa-rules/tsconfig.json
@@ -12,6 +12,7 @@
     "src/common/group.ts",
     "src/common/outcome/contrast.ts",
     "src/common/outcome/text-spacing.ts",
+    "src/common/predicate.ts",
     "src/common/predicate/has-accessible-name.ts",
     "src/common/predicate/has-attribute.ts",
     "src/common/predicate/has-cascaded-value-declared-in-inline-style.ts",
@@ -36,6 +37,7 @@
     "src/common/predicate/is-ignored.ts",
     "src/common/predicate/is-inert.ts",
     "src/common/predicate/is-keyboard-actionable.ts",
+    "src/common/predicate/is-large-text.ts",
     "src/common/predicate/is-marked-decorative.ts",
     "src/common/predicate/is-perceivable.ts",
     "src/common/predicate/is-rendered.ts",
